slabs: refactor downloadShards
